sales_leads_notifier/requirements.txt
google-cloud-firestore==2.16.0
google-cloud-aiplatform==1.24.0
google-auth==2.17.3
python-dotenv==1.0.0
requests==2.31.0